requirements.txt
flask
fpdf2
dropbox-sign
python-dotenv
flask-cors
